pyproject.toml
[project]
name = "mockup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=44.0.0",
    "fastapi>=0.115.6",
    "itsdangerous>=2.2.0",
    "passlib>=1.7.4",
    "psycopg2>=2.9.10",
    "pydantic[email]>=2.10.4",
    "python-dotenv>=1.0.1",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.36",
    "uvicorn>=0.34.0",
]
